requirements.txt
django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.10.0
python-decouple==3.4
gunicorn==20.1.0
